% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stan.model}
\alias{stan.model}
\title{A fitted Bayesian model.}
\format{
A list:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
stan.model
}
\description{
A dataset containing the fitted Bayesian model.
}
\keyword{datasets}
